Changes necessary to compile on Windows with MSVC
Mostly involved rigging up export symbols for the shared library. On
Windows symbols are not exported by default (unlike Linux). In the case
where you have a shared library with no exported symbols Windows does
not create a .lib file (which I am guessing contains the information
about these exported symbols) and you end up with a somewhat cryptic
cmake error when you try to build the test runner which depends on it.
Luckily, cmake provides the machinery to automatically generate this
export symbol header via the GenerateExportHeader function/macro and the
rest is just making sure the include paths are set up properly and the
export symbols are in the right places.